Add timestamp to gracefully remove stale transactions
There are instances where a stale transaction stays around do to an error in the mutation. As a result, the recursive update_transaction_queue method will continue to run over and over, causing an inifinte loop and potentially crashing the server.

This attempts to add a timestamp once the transaction is validated when checking via the timestamp and will remove it from the queue.